feat(feedback): Hide the URL section for mobile projects
Fixes REPLAY-454
